fix: broken url in feed on mobile
