+++ packages/app/components/swipe-list.web.tsx
@@ -1,11 +1,4 @@
-import {
-  useCallback,
-  useMemo,
-  useRef,
-  useEffect,
-  createContext,
-  useState,
-} from "react";
+import { useCallback, useMemo, useRef, createContext, useState } from "react";
 import { useWindowDimensions } from "react-native";
 
 import { useSharedValue } from "react-native-reanimated";
@@ -24,9 +17,7 @@ import {
   ViewabilityItemsContext,
 } from "app/components/viewability-tracker-flatlist";
 import { VideoConfigContext } from "app/context/video-config-context";
-import { getNFTSlug } from "app/hooks/use-share-nft";
 import { useScrollToTop } from "app/lib/react-navigation/native";
-import { createParam } from "app/navigation/use-param";
 import type { NFT } from "app/types";
 import { isMobileWeb } from "app/utilities";
 
@@ -37,23 +28,22 @@ type Props = {
   refresh?: () => void;
   initialScrollIndex?: number;
   bottomPadding?: number;
+  onVisibleIndexChange?: (index: number) => void;
 };
-const { useParam } = createParam();
 
 export const SwiperActiveIndexContext = createContext<number>(0);
 export const SwipeList = ({
   data,
   fetchMore,
   initialScrollIndex = 0,
+  onVisibleIndexChange,
 }: Props) => {
   // Todo: use nft_id instead of initialScrollIndex navigate to specific NFT
   // const [id, setId] = useParam("id");
-  const [, setInitialScrollIndex] = useParam("initialScrollIndex");
 
   const [activeIndex, setActiveIndex] = useState(0);
   const listRef = useRef<any>(null);
   useScrollToTop(listRef);
-  const initialURLSet = useRef(false);
 
   const visibleItems = useSharedValue<any[]>([
     undefined,
@@ -70,12 +60,6 @@ export const SwipeList = ({
     []
   );
 
-  useEffect(() => {
-    if (!initialURLSet.current) {
-      window.history.replaceState(null, "", getNFTSlug(data[0]));
-      initialURLSet.current = true;
-    }
-  }, [data]);
   // const initialSlideIndex = useMemo(() => {
   //   const defaultIndex = clamp(initialScrollIndex, 0, data.length - 1);
   //   if (!id) return defaultIndex;
@@ -90,10 +74,10 @@ export const SwipeList = ({
         e.activeIndex,
         e.activeIndex + 1 < data.length ? e.activeIndex + 1 : undefined,
       ];
-      window.history.replaceState(null, "", getNFTSlug(data[e.activeIndex]));
+      onVisibleIndexChange?.(e.activeIndex);
       setActiveIndex(e.activeIndex);
     },
-    [visibleItems, data]
+    [visibleItems, data, onVisibleIndexChange]
   );
 
   if (data.length === 0) return null;
