+++ packages/app/screens/swipe-list.tsx
@@ -1,4 +1,5 @@
-import { useMemo } from "react";
+import { useMemo, useEffect, useRef, useCallback } from "react";
+import { Platform } from "react-native";
 
 import { useSafeAreaInsets } from "@showtime-xyz/universal.safe-area";
 
@@ -8,6 +9,7 @@ import { ProfileTabsNFTProvider } from "app/context/profile-tabs-nft-context";
 import { useTrendingCreators, useTrendingNFTS } from "app/hooks/api-hooks";
 import { useProfileNFTs } from "app/hooks/api-hooks";
 import { useFeed } from "app/hooks/use-feed";
+import { getNFTSlug } from "app/hooks/use-share-nft";
 import { useUser } from "app/hooks/use-user";
 import { useTrackPageViewed } from "app/lib/analytics";
 import { createParam } from "app/navigation/use-param";
@@ -52,10 +54,12 @@ const FeedSwipeList = () => {
   const { data } = useFeed();
   const [initialScrollIndex] = useParam("initialScrollIndex");
   const { bottom: safeAreaBottom } = useSafeAreaInsets();
+  const { onVisibleIndexChange } = useSyncNFTURL(data);
 
   return (
     <SwipeList
       data={data}
+      onVisibleIndexChange={onVisibleIndexChange}
       initialScrollIndex={Number(initialScrollIndex)}
       bottomPadding={safeAreaBottom}
     />
@@ -80,6 +84,7 @@ const ProfileSwipeList = () => {
     }
   );
   const { bottom: safeAreaBottom } = useSafeAreaInsets();
+  const { onVisibleIndexChange } = useSyncNFTURL(data);
 
   return (
     <MutateProvider mutate={updateItem}>
@@ -92,6 +97,7 @@ const ProfileSwipeList = () => {
       >
         <SwipeList
           data={data}
+          onVisibleIndexChange={onVisibleIndexChange}
           fetchMore={fetchMore}
           isRefreshing={isRefreshing}
           refresh={refresh}
@@ -112,6 +118,7 @@ const TrendingNFTsSwipeList = () => {
     days: Number(days),
   });
   const { bottom: safeAreaBottom } = useSafeAreaInsets();
+  const { onVisibleIndexChange } = useSyncNFTURL(data);
 
   return (
     <SwipeList
@@ -119,6 +126,7 @@ const TrendingNFTsSwipeList = () => {
       // fetchMore={fetchMore}
       // isRefreshing={isRefreshing}
       // refresh={refresh}
+      onVisibleIndexChange={onVisibleIndexChange}
       initialScrollIndex={Number(initialScrollIndex)}
       bottomPadding={safeAreaBottom}
     />
@@ -130,6 +138,7 @@ export const TrendingCreatorSwipeList = withColorScheme(() => {
   const [days] = useParam("days");
   const [initialScrollIndex] = useParam("initialScrollIndex");
   const [creatorId] = useParam("creatorId");
+  const { onVisibleIndexChange } = useSyncNFTURL(data);
 
   const { data, mutate } = useTrendingCreators({
     days: Number(days),
@@ -160,7 +169,32 @@ export const TrendingCreatorSwipeList = withColorScheme(() => {
         fetchMore={() => {}}
         isRefreshing={false}
         refresh={() => {}}
+        onVisibleIndexChange={onVisibleIndexChange}
       />
     </MutateProvider>
   );
 });
+
+const useSyncNFTURL = (data: NFT[]) => {
+  const initialURLSet = useRef(false);
+
+  useEffect(() => {
+    if (Platform.OS !== "web") return;
+
+    if (!initialURLSet.current) {
+      window.history.replaceState(null, "", getNFTSlug(data[0]));
+      initialURLSet.current = true;
+    }
+  }, [data]);
+
+  const onVisibleIndexChange = useCallback(
+    (index: number) => {
+      if (Platform.OS !== "web") return;
+
+      window.history.replaceState(null, "", getNFTSlug(data[index]));
+    },
+    [data]
+  );
+
+  return { onVisibleIndexChange };
+};
